backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.0.19
pydantic==2.0.3
python-dotenv==1.0.0
httpx==0.24.1
beautifulsoup4==4.12.2
requests==2.31.0